main.py
"""献立表カレンダー作成ゲーム - エントリポイント"""

import sys
import pygame
from src.constants import SCREEN_WIDTH, SCREEN_HEIGHT, FPS, TITLE, COLOR_BG


def main():
    pygame.init()
    screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT))
    pygame.display.set_caption(TITLE)
    clock = pygame.time.Clock()

    running = True
    while running:
        for event in pygame.event.get():
            if event.type == pygame.QUIT:
                running = False

        screen.fill(COLOR_BG)
        pygame.display.flip()
        clock.tick(FPS)

    pygame.quit()
    sys.exit()


if __name__ == "__main__":
    main()
