requirements.txt
pygame-ce>=2.5.0
ortools>=9.8
pytest>=8.0
